Trying to get modal working
